Split GameHelper into smaller modules [COHESION]
