+++ src/components/Card.vue
@@ -9,8 +9,6 @@
 </template>
 
 <script>
-// import { mapActions } from 'vuex'
-
 export default {
   props: {
     card: {
