+++ src/js/PlayerHelper.js
@@ -1,0 +1,19 @@
+import { LOCATION } from '../js/DeckHelper'
+
+export const PLAYER_TYPE = {
+  HUMAN: 'human',
+  CPU: 'cpu'
+}
+
+export const PLAYER_ROLE = {
+  PERSONALITY1: 'cpu1',
+  PERSONALITY2: 'cpu2',
+  PERSONALITY3: 'cpu3'
+}
+
+export const PLAYER_LOCATION = [
+  LOCATION.PLAYER1,
+  LOCATION.PLAYER2,
+  LOCATION.PLAYER3,
+  LOCATION.PLAYER4
+]
