+++ src/pages/Multi.vue
@@ -22,7 +22,7 @@
 <script>
 import { mapState, mapGetters, mapActions } from 'vuex'
 import { fdb } from '../vuex/firebase'
-import { PLAYER_TYPE, PLAYER_ROLE } from '../js/GameHelper'
+import { PLAYER_TYPE, PLAYER_ROLE } from '../js/PlayerHelper'
 import Board from 'components/Board'
 
 export default {
